Create base structure for executors
The structure is:
  - Conduit Core Application
  - Conduit Supervisor
  - Conduit Executor

The idea will then be to have an interactor call the correct module
and in term the correct service. If it crashes there will be no problem
since we're using :one_to_one recovery.

Finally, when we're ready, we'll start adding evaluators and possibly
change the structure to use protocols instead of interactors to
recognize the correct process to call.