Change to :simple_one_for_one strategy
Changed the Supervisor strategy so we can create executors on demand
to have more control over running processes and to scale it using
more children instead of spawns.

Due to the new nature of this, we don't want to handle cast since we
can have that use case by spawning children.