Create base structure for executors (#1)
* Create base structure for executors

The structure is:
  - Conduit Core Application
  - Conduit Supervisor
  - Conduit Executor

The idea will then be to have an interactor call the correct module
and in term the correct service. If it crashes there will be no problem
since we're using :one_to_one recovery.

Finally, when we're ready, we'll start adding evaluators and possibly
change the structure to use protocols instead of interactors to
recognize the correct process to call.

* Add Elixir Evaluation

Added String Eval to Elixir Executor.

This kind of shows the basic idea, receive a bunch of lines,
eval and compose them into a single function and apply a simple
input (lists, maps, sets, etc...).

Example of this:

`ConduitCore.Executor.Elixir.execute_async(
  [
    "fn (l) -> Enum.reverse(l) end",
    "&(IO.inspect/1)"
  ], [1,2,3])`

* Fix nameless supervisor issues

Currently the supervisor had no name which created some issues when
we wanted to list children.

Now that is fixed so it will be easier to call the correct child
when required.

* Change to :simple_one_for_one strategy

Changed the Supervisor strategy so we can create executors on demand
to have more control over running processes and to scale it using
more children instead of spawns.

Due to the new nature of this, we don't want to handle cast since we
can have that use case by spawning children.

* Change the structure of the Executors

Since we made the changes to spawn a child for each execution, we now
need to have a Supervisor per language. As such we started to supervise
the supervisors!

* Add better logging

Added better logging and avoided the usage of puts and inspect